backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.2.2
Pillow==9.0.0
webcolors==1.11.1
django-filter==21.1
psycopg2-binary==2.9.3
flake8==6.0.0
flake8-isort==6.0.0
